+++ src/ext/java/org/opentripplanner/ext/smoovebikerental/SmooveBikeRentalDataSource.java
@@ -13,21 +13,26 @@ import org.slf4j.LoggerFactory;
 
 /**
  * Implementation of a VehicleRentalDataSource for the Smoove GIR SabiWeb used in Helsinki.
+ *
  * @see VehicleRentalDataSource
  */
-public class SmooveBikeRentalDataSource extends GenericJsonVehicleRentalDataSource<SmooveBikeRentalDataSourceParameters> {
+public class SmooveBikeRentalDataSource
+        extends GenericJsonVehicleRentalDataSource<SmooveBikeRentalDataSourceParameters> {
 
     private static final Logger log = LoggerFactory.getLogger(SmooveBikeRentalDataSource.class);
 
     public static final String DEFAULT_NETWORK_NAME = "smoove";
 
+    private final boolean allowOverloading;
+
     private final String networkName;
     private final RentalVehicleType vehicleType;
 
     public SmooveBikeRentalDataSource(SmooveBikeRentalDataSourceParameters config) {
-        super(config,"result");
+        super(config, "result");
         networkName = config.getNetwork(DEFAULT_NETWORK_NAME);
         vehicleType = RentalVehicleType.getDefaultType(networkName);
+        allowOverloading = config.isAllowOverloading();
     }
 
     /**
@@ -63,14 +68,19 @@ public class SmooveBikeRentalDataSource extends GenericJsonVehicleRentalDataSour
             return null;
         }
         if (!node.path("operative").asText().equals("true")) {
+            station.isRenting = false;
+            station.isReturning = false;
             station.vehiclesAvailable = 0;
             station.spacesAvailable = 0;
+            station.capacity = node.path("total_slots").asInt();
         } else {
             station.vehiclesAvailable = node.path("avl_bikes").asInt();
             station.spacesAvailable = node.path("free_slots").asInt();
+            station.capacity = node.path("total_slots").asInt();
         }
-        station.vehicleTypesAvailable = Map.of(vehicleType, station.spacesAvailable);
+        station.vehicleTypesAvailable = Map.of(vehicleType, station.vehiclesAvailable);
         station.vehicleSpacesAvailable = Map.of(vehicleType, station.spacesAvailable);
+        station.allowOverloading = allowOverloading;
         return station;
     }
 }