+++ src/main/java/org/opentripplanner/api/parameter/ApiRequestMode.java
@@ -6,7 +6,7 @@ public enum ApiRequestMode {
     WALK, BICYCLE, SCOOTER, CAR,
     TRAM, SUBWAY, RAIL, BUS, COACH, FERRY,
     CABLE_CAR, GONDOLA, FUNICULAR,
-    TRANSIT, AIRPLANE, FLEX;
+    TRANSIT, AIRPLANE, FLEX, TROLLEYBUS;
 
     public static ApiRequestMode fromTransitMode(TransitMode transitMode) {
         switch (transitMode) {
@@ -30,6 +30,8 @@ public enum ApiRequestMode {
                 return GONDOLA;
             case FUNICULAR:
                 return FUNICULAR;
+            case TROLLEYBUS:
+                return TROLLEYBUS;
             default:
                 throw new IllegalArgumentException("Can't convert to ApiRequestMode: " + transitMode);
         }