+++ src/main/java/org/opentripplanner/api/parameter/QualifiedModeSet.java
@@ -74,6 +74,9 @@ public class QualifiedModeSet implements Serializable {
                  case FUNICULAR:
                      transitModes.add(TransitMode.FUNICULAR);
                      break;
+                 case TROLLEYBUS:
+                     transitModes.add(TransitMode.TROLLEYBUS);
+                     break;
              }
         }