+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/transit/request/RaptorRoutingRequestTransitData.java
@@ -13,6 +13,7 @@ import org.opentripplanner.model.transfer.TransferService;
 import org.opentripplanner.routing.algorithm.raptor.transit.TransitLayer;
 import org.opentripplanner.routing.algorithm.raptor.transit.TripSchedule;
 import org.opentripplanner.routing.algorithm.raptor.transit.cost.DefaultCostCalculator;
+import org.opentripplanner.routing.algorithm.raptor.transit.mappers.DateMapper;
 import org.opentripplanner.routing.algorithm.raptor.transit.mappers.McCostParamsMapper;
 import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.transit.raptor.api.transit.CostCalculator;
@@ -51,11 +52,15 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
 
   private final CostCalculator generalizedCostCalculator;
 
+  private final int validTransitDataStartTime;
+
+  private final int validTransitDataEndTime;
 
   public RaptorRoutingRequestTransitData(
       TransferService transferService,
       TransitLayer transitLayer,
       Instant departureTime,
+      int additionalPastSearchDays,
       int additionalFutureSearchDays,
       TransitDataProviderFilter filter,
       RoutingRequest routingRequest
@@ -72,6 +77,7 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
     );
     this.startOfTime = transitDataCreator.getSearchStartTime();
     this.activeTripPatternsPerStop = transitDataCreator.createTripPatternsPerStop(
+        additionalPastSearchDays,
         additionalFutureSearchDays,
         filter
     );
@@ -79,6 +85,15 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
     this.generalizedCostCalculator = new DefaultCostCalculator(
             McCostParamsMapper.map(routingRequest),
             transitLayer.getStopIndex().stopBoardAlightCosts
+    );
+    this.validTransitDataStartTime = DateMapper.secondsSinceStartOfTime(
+        startOfTime,
+        startOfTime.minusDays(additionalPastSearchDays).toInstant()
+    );
+    // The +1 is due to the validity being to the end of the day
+    this.validTransitDataEndTime = DateMapper.secondsSinceStartOfTime(
+        startOfTime,
+        startOfTime.plusDays(additionalFutureSearchDays + 1).toInstant()
     );
   }
 
@@ -139,6 +154,16 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
       var s = transitLayer.getStopByIndex(stopIndex);
       return s==null ? "null" : s.getName() + "(" + stopIndex + ")";
     };
+  }
+
+  @Override
+  public int getValidTransitDataStartTime() {
+    return validTransitDataStartTime;
+  }
+
+  @Override
+  public int getValidTransitDataEndTime() {
+    return validTransitDataEndTime;
   }
 
   public ZonedDateTime getStartOfTime() {