+++ src/main/java/org/opentripplanner/routing/core/TraverseMode.java
@@ -8,10 +8,10 @@ public enum TraverseMode {
     WALK, BICYCLE, SCOOTER, CAR,
     TRAM, SUBWAY, RAIL, BUS, FERRY,
     CABLE_CAR, GONDOLA, FUNICULAR,
-    TRANSIT, AIRPLANE;
+    TRANSIT, AIRPLANE, TROLLEYBUS, MONORAIL;
 
     private static final EnumSet<TraverseMode> TRANSIT_MODES = EnumSet.of(
-        AIRPLANE, BUS, CABLE_CAR, FERRY, FUNICULAR, GONDOLA, RAIL, SUBWAY, TRAM, TRANSIT
+        AIRPLANE, BUS, CABLE_CAR, FERRY, FUNICULAR, GONDOLA, RAIL, SUBWAY, TRAM, TRANSIT, TROLLEYBUS, MONORAIL
     );
 
     private static final EnumSet<TraverseMode> STREET_MODES = EnumSet.of(WALK, BICYCLE, SCOOTER, CAR);
@@ -39,6 +39,7 @@ public enum TraverseMode {
     public static TraverseMode fromTransitMode(TransitMode transitMode) {
         switch (transitMode) {
             case RAIL:
+            case MONORAIL:
                 return TraverseMode.RAIL;
             case COACH:
             case BUS:
@@ -57,6 +58,8 @@ public enum TraverseMode {
                 return TraverseMode.GONDOLA;
             case FUNICULAR:
                 return TraverseMode.FUNICULAR;
+            case TROLLEYBUS:
+                return TraverseMode.TROLLEYBUS;
             default:
                 throw new IllegalArgumentException();
         }